# Optimize ResourceManager