--- a/src/OrchardCore/OrchardCore.ResourceManagement.Abstractions/RequireSettings.cs
+++ b/src/OrchardCore/OrchardCore.ResourceManagement.Abstractions/RequireSettings.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Linq;
 using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 
 namespace OrchardCore.ResourceManagement
 {
@@ -20,9 +21,10 @@ namespace OrchardCore.ResourceManagement
         public string Version { get; set; }
         public bool? AppendVersion { get; set; }
         public Action<ResourceDefinition> InlineDefinition { get; set; }
+
         public Dictionary<string, string> Attributes
         {
-            get { return _attributes ?? (_attributes = new Dictionary<string, string>()); }
+            get => _attributes ?? (_attributes = new Dictionary<string, string>());
             private set { _attributes = value; }
         }
 
@@ -48,6 +50,7 @@ namespace OrchardCore.ResourceManagement
         /// <summary>
         /// The resource will be displayed in the head of the page
         /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings AtHead()
         {
             return AtLocation(ResourceLocation.Head);
@@ -57,6 +60,7 @@ namespace OrchardCore.ResourceManagement
         /// The resource will be displayed at the foot of the page
         /// </summary>
         /// <returns></returns>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings AtFoot()
         {
             return AtLocation(ResourceLocation.Foot);
@@ -66,6 +70,7 @@ namespace OrchardCore.ResourceManagement
         /// The resource will be displayed at the specified location
         /// </summary>
         /// <param name="location">The location where the resource should be displayed</param>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings AtLocation(ResourceLocation location)
         {
             // if head is specified it takes precedence since it's safer than foot
@@ -73,6 +78,7 @@ namespace OrchardCore.ResourceManagement
             return this;
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings UseCulture(string cultureName)
         {
             if (!String.IsNullOrEmpty(cultureName))
@@ -82,46 +88,54 @@ namespace OrchardCore.ResourceManagement
             return this;
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings UseDebugMode()
         {
             return UseDebugMode(true);
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings UseDebugMode(bool debugMode)
         {
             DebugMode |= debugMode;
             return this;
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings UseCdn()
         {
             return UseCdn(true);
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings UseCdn(bool cdn)
         {
             CdnMode |= cdn;
             return this;
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings UseCdnBaseUrl(string cdnBaseUrl)
         {
             CdnBaseUrl = cdnBaseUrl;
             return this;
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings WithBasePath(string basePath)
         {
             BasePath = basePath;
             return this;
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings UseCondition(string condition)
         {
             Condition = Condition ?? condition;
             return this;
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings UseVersion(string version)
         {
             if (!String.IsNullOrEmpty(version))
@@ -131,6 +145,7 @@ namespace OrchardCore.ResourceManagement
             return this;
         }
 
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public RequireSettings ShouldAppendVersion(bool? appendVersion)
         {
             AppendVersion = appendVersion;