--- a/test/OrchardCore.Benchmarks/Support/FakeFileProvider.cs
+++ b/test/OrchardCore.Benchmarks/Support/FakeFileProvider.cs
@@ -1,0 +1,25 @@
+using System.Threading;
+using Microsoft.Extensions.FileProviders;
+using Microsoft.Extensions.Primitives;
+using OrchardCore.DisplayManagement.FileProviders;
+
+namespace OrchardCore.Benchmark.Support
+{
+    internal class FakeFileProvider : IFileProvider
+    {
+        public IDirectoryContents GetDirectoryContents(string subpath)
+        {
+            return new NotFoundDirectoryContents();
+        }
+
+        public IFileInfo GetFileInfo(string subpath)
+        {
+            return new ContentFileInfo("name", "content");
+        }
+
+        public IChangeToken Watch(string filter)
+        {
+            return new CancellationChangeToken(CancellationToken.None);
+        }
+    }
+}