This starts to take shape, don't know how many bugs I've introduced but biggest offenders should be handled now.

``` ini
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.18980
Intel Core i7-6820HQ CPU 2.70GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
.NET Core SDK=3.0.100-preview9-014004
  [Host]     : .NET Core 3.0.0-preview9-19423-09 (CoreCLR 4.700.19.42102, CoreFX 4.700.19.42104), 64bit RyuJIT
  DefaultJob : .NET Core 3.0.0-preview9-19423-09 (CoreCLR 4.700.19.42102, CoreFX 4.700.19.42104), 64bit RyuJIT
```

## Before

|           Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| RenderStylesheet | 18.03 us | 0.0706 us | 0.0589 us | 3.6926 |     - |     - |  15.16 KB |

## After

|           Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
| RenderStylesheet | 9.379 us | 0.0670 us | 0.0627 us | 1.9836 |     - |     - |   8.16 KB |