@lahma 2019-09-16
I think biggest offenders have been taken care of, so merge is fine. There's little test coverage so hopefully I didn't break it badly...